requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

gradio==4.13.0
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.8
more-itertools==10.1.0
numba==0.58.1
numpy==1.26.3
pandas==2.1.4
pyspellchecker==0.7.3
regex==2023.12.25
scipy==1.11.4
setuptools==58.0.4
tiktoken==0.5.2
torch==2.1.2
tqdm==4.66.1
